+++ src/components/ColorAndIconPicker/IconPickerModal.tsx
@@ -2,6 +2,7 @@ import i18n from '@dhis2/d2-i18n'
 import {
     Button,
     ButtonStrip,
+    CircularLoader,
     Input,
     Modal,
     ModalActions,
@@ -11,8 +12,9 @@ import {
     TabBar,
 } from '@dhis2/ui'
 import cx from 'classnames'
-import React, { useState } from 'react'
-import { filterIcons } from './filterIcons'
+import React, { useEffect, useMemo, useState } from 'react'
+import { useDebouncedState } from '../../lib'
+import { LoadingSpinner } from '../loading/LoadingSpinner'
 import classes from './IconPickerModal.module.css'
 import { useIconsQuery, Icon } from './useIconsQuery'
 
@@ -27,13 +29,49 @@ export function IconPickerModal({
     onChange: ({ icon }: { icon: string }) => void
     onCancel: () => void
 }) {
-    const [searchValue, setSearchValue] = useState('')
+    const {
+        liveValue: searchValue,
+        setValue: setSearchValue,
+        debouncedValue: debouncedSearchValue,
+    } = useDebouncedState({
+        initialValue: '',
+    })
+    const [loadingRef, setLoadingRef] = useState<HTMLDivElement | null>(null)
     const [activeTab, setActiveTab] = useState<TabName>('all')
+
     const [icon, setIcon] = useState(selected)
-    const icons = useIconsQuery()
-    const displayIcons = searchValue
-        ? filterIcons(icons.data[activeTab], searchValue)
-        : icons.data[activeTab]
+    const icons = useIconsQuery({
+        search: debouncedSearchValue.trim(),
+    })
+
+    const fetchNextPage = icons.fetchNextPage
+
+    useEffect(() => {
+        if (loadingRef) {
+            const observer = new IntersectionObserver(
+                (entries) => {
+                    if (entries[0].isIntersecting) {
+                        fetchNextPage()
+                    }
+                },
+                {
+                    threshold: 1,
+                }
+            )
+            observer.observe(loadingRef)
+            return () => {
+                observer.disconnect()
+            }
+        }
+    }, [fetchNextPage, loadingRef])
+
+    const displayIcons = useMemo(
+        () =>
+            activeTab !== 'all'
+                ? icons.allData.filter((i) => i.key.endsWith(activeTab))
+                : icons.allData,
+        [icons.allData, activeTab]
+    )
 
     return (
         <Modal large onClose={onCancel}>
@@ -46,28 +84,28 @@ export function IconPickerModal({
                             onClick={() => setActiveTab('all')}
                             selected={activeTab === 'all'}
                         >
-                            All
+                            {i18n.t('All')}
                         </Tab>
 
                         <Tab
                             onClick={() => setActiveTab('positive')}
                             selected={activeTab === 'positive'}
                         >
-                            Positive
+                            {i18n.t('Positive')}
                         </Tab>
 
                         <Tab
                             onClick={() => setActiveTab('negative')}
                             selected={activeTab === 'negative'}
                         >
-                            Negative
+                            {i18n.t('Negative')}
                         </Tab>
 
                         <Tab
                             onClick={() => setActiveTab('outline')}
                             selected={activeTab === 'outline'}
                         >
-                            Outline
+                            {i18n.t('Outline')}
                         </Tab>
                     </TabBar>
 
@@ -79,26 +117,50 @@ export function IconPickerModal({
                             onChange={({ value }) =>
                                 setSearchValue(value || '')
                             }
+                            type="search"
                         />
                     </div>
                 </div>
 
-                <div className={classes.iconsContainer}>
-                    {displayIcons.map(({ key, description, href }: Icon) => (
+                <div
+                    className={classes.iconsContainer}
+                    data-test="icons-container"
+                >
+                    <div className={classes.iconMessage}>
+                        {icons.isLoading && <LoadingSpinner />}
+                        {icons.isSuccess && displayIcons.length === 0 ? (
+                            <span>{i18n.t('No icons match the search.')}</span>
+                        ) : null}
+                    </div>
+                    {displayIcons.map(({ key, href }: Icon) => (
                         <div
                             key={key}
                             className={cx(classes.iconContainer, {
                                 [classes.active]: key === icon,
                             })}
                             onClick={() => setIcon(key)}
+                            title={key}
                         >
                             <img
                                 className={classes.iconImage}
-                                alt={description}
+                                alt={key}
                                 src={href}
+                                loading="lazy"
                             />
                         </div>
                     ))}
+                    {icons.hasNextPage && !icons.isFetching && (
+                        <div
+                            className={classes.iconsLoading}
+                            ref={(ref) => {
+                                if (!!ref && ref !== loadingRef) {
+                                    setLoadingRef(ref)
+                                }
+                            }}
+                        >
+                            <CircularLoader />
+                        </div>
+                    )}
                 </div>
             </ModalContent>